support emptying groups and resetting w/o clearing.
in other news, here's an interesting way to think about this little
demo:

we're assigning attributes to a collection of dots. then we group them
by those attributes and perform actions (drawing lines) based on the
groups.
the data (the dots) is separate from the meaning we assign to them. it's
just a little visualization, with some interpretation (order of dots to
connect) thrown in.

pixl is also a dot-drawing applications, with restrictions on where you
can draw the dots.

what's the pattern?