(Try to) make sentence highlighting more robust
The previous highlighting didn't search for the closest match, it
stopped when it found one, even if there were other closer ones.  We do
that now, at the cost of (almost) duplicated code (`lastIndexOf` vs
`indexOf` and `>` vs `<`).  However, it's still imperfect.  For some
reason the highlighting doesn't work until you're two characters "into"
a sentence, at least if you're at the end of one.  Also, highlighting
when no sentence is clicked is weird.  Quite a bit of work to do.